another way to fix GRPC dependency
Summary:
See the full summary of the issue in D15591571.

In this diff, instead of running the complete post-install script of `node-pre-gyp`, we skip all of that and simply extract the binary we know this module requires in the correct directory.

This removes the dependency on `node-pre-gyp` and its build process, but also makes this prone to breaking easily since versions are hardcoded and we will need to go update those if anything is updated.

Reviewed By: rckclmbr

Differential Revision: D15600238

fbshipit-source-id: 2b16af3771757dcac3f0d57357c22dd37407bc9a