Add sctpd service to EPC (#320)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/320

As part of the effort to make MME stateless, we need a way to preserve state across MME restarts for each subcomponent. This includes SCTP. Unfortunately, the kernel doesn't offer any facility to keep an SCTP socket alive across process restarts. Therefore, we are factoring out MME SCTP facilities into a separate, more stable service, sctpd.

## [magma][sctpd] Add classes for sctp context and associations
Since we are factoring out sctp from MME into a C++ based service, we will do SCTP bookkeeping in a C++ manner (more clear/type safe).

## [magma][sctpd] Add util file for temporary logging and sctp config
To support sctpd, we need to have basic logging facilities and sctp socket manipulation. Use of MagmaLogging is planned for the future but is not used right now (#movefast).

## [magma][sctpd] Add SctpUplink client stubs
Here, we add the sctpd uplink grpc client so that the sctpd can communicate to MME: upstream data packets, new associations and closed/reset associations.

## [magma][sctpd] Add sctp listener thread implemention for upstream tx
Here, we add a listener thread implementation to wait for new associations and data packets available on an sctp socket. This uses the previous stcpd uplink client to relay these events to the MME.

## [magma][sctpd] Add SctpDownlink rpc server implemention for downstream tx
Here, we add the sctpd downlink server implementation. This listens for communications from MME requesting for an sctp connection to setup or to send a downlink packet (mme -> enb) on an existing connection.

The downlink server will be responsible for spawning the sctp listener thread (upstream) after an initialization request is received from the mme.

## [magma][sctpd] Add main sctpd file and associated build files
Here, we add the main sctpd file. This is in charge of launching the downlink server (which in turn waits for MME connection).

Additionally, we add the files necessary to build sctpd (CMakeLists.txt and changes to Makefile)

## [magma][mme] Make log.h/c compatible with cpp
Currently, including log.h from a C++ file generates compile errors due to C++ being more strict with char * vs const char *. This diff fixes log.h to be C++ compatible

## [magma][mme] Add appropriate includes to sctp_messages_types.h
Currently, sctp_messages_types.h doesn't include the proper types to compile on its own.
Let's add those.

## [magma][mme] Change sctp task to use sctpd
Here, we modify MME's sctp task to defer to sctpd for all sctp related communications instead of locally managing the sctp sockets.
Note: at this point in time sctpd is effectively stateless. This does not play well with the stateless-less MME across restarts. This will be handled better in a following (stacked) diff.

## [magma][sctpd] Move socket fd into SctpDesc
The socket descriptor is a logical part of an SCTP description (state). Let's put it there.

## [magma][sctpd] Class-ify sctp_listener to make shutdown easier
Currently, communicating with the listener thread is limited to pass parameters to the thread construction. Since, the handle returned is just the std::thread object, don't have easy insight or control over the threads behavior. Let's put the listener thread into a controlling class with an interface similar to a grpc server (Start, Stop, Wait methods).

## [magma][sctpd] Add force restart option to sctpd for stateless-less deployment
Right now, sctpd is stateless while MME is stateless-less. This doesn't work out well when MME restarts unexpectedly as there is residual state in sctpd not reflected in MME. This diff adds a force_restart option to sctpd.init so that MME can request sctpd to open a new connection, therefore resetting the state.

## [magma][sctpd] Configure socket options to fail hard
Currently, we encounter an error in s1ap tester where an sctp socket is opened successfully, but s1ap tester is unable to communicate with the eNB. The likely cause is s1ap tester not closing the socket properly after failing a test. Additionally, we have SO_REUSEADDR set on the socket which allows us to open another socket even if the other has not closed properly disguising the issue.
Let's remove the SO_REUSEADDR socket option so that we fail hard whenever this is the case.

## [magma][sctpd] Update sctpd to use references instead of shared_ptr's
In this diff we change various uses of shared_ptr to refs to be more in line with C++ style. Additionally, we mark certain methods as const to show they have no side effects.

## [magma][sctpd] Add comments to classes/methods exposed in header files
Previously, sctpd didn't have comments in file headers. Let's add those so the code is better documented.

## [magma][sctpd] Use glog/magma_logging instead of XLOG
Previously, sctpd used printf for logging. Let's change it to use magma_logging (and therefore GLOG) instead to get more robust/controllable logging behavior.

## [magma][sctpd] Apply clang format to sctpd
Some of the past diffs for sctpd have had formatting errors. Rather than fix them in each diff and deal with rebasing issues let's fix them all at once with clang-format

## [magma][sctpd] Add sctpd setup to provision setup and systemd configs
Previously, we had to manually run sctpd to test and run it as part of EPC. Let's add sctpd to the configuration files so that it is run automatically by magmad/make start.

Reviewed By: ssanadhya

Differential Revision: D15090188

fbshipit-source-id: 746ad497d3668daeef08df8d55e74621311e591d