Fix magma_conditional_service
Summary:
This diff follows https://our.intern.facebook.com/intern/diff/D15562556/
to try and restore functionality.

In the above diff, getting mcconfigs as structures is mostly deprecated, instead
relying mostly on returning dicts.

This diff attempts to patch things over.

The main change is that the returned value is a dict. I originally had convinced
myself that getting unset values from the structure would be different than
getting unset values from the dict, but don't think that's actually true in
practice. I left those in for your consideration, but the minimum form of
this diff is a 1-liner.

Also randomly delint.

Reviewed By: xjtian

Differential Revision: D15739346

fbshipit-source-id: a41e6feaa320dde46c6188e067bb260880b2d877