Remove daemon creation support
Summary:
Currently, we use systemd for managing MME and other services on AGW. This does not rely on the process daemonizing itself (and we do not).

Let's remove the daemonization code as long as we are not using the functionality.

NOTE: If you think we still need this for something in the foreseeable future let me know. However, I the daemonization code is small and should be easy to add back.

Reviewed By: ssanadhya

Differential Revision: D15343663

fbshipit-source-id: fd1a250dbec6061afcf4b2d46f70abe81aded882