Schedule revalidation based on the information in credit update response (#333)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/333

**Summary**
Both RAR and update response message could contain revalidation timeout which is a type of event triggers that requires `sessiond` to send an update after the timer expires. This diff implements the logic required in handling update response message.

**Implementation**
1. Check whether revalidation timer is presented in the update response message and schedule an update if it is required.
2. Update related unit tests.

**What is affected**
`Sessiond` should schedule and send an update if revalidation timeout is one of the event triggers in update response message.
Nothing else should be affected.

Reviewed By: emakeev

Differential Revision: D15770428

fbshipit-source-id: d401d185043cd7a5f77039d2a1996c4644e657ce