Add mock task implementation for fuzzing
Summary:
In order to improve the security and stability of the MME, it would be helpful to add fuzz testing. However, to do this we need tasks to be able to run independently of one another.

This adds mock tasks -- used to fuzz tasks independent of one another. These mock tasks can be brought up by the main thread and their messages queues synchronously read.

Reviewed By: ssanadhya

Differential Revision: D14171469

fbshipit-source-id: 79b7c98fa6aef2e8ec97cac944fd5c160fb08016