Add example fuzz target
Summary:
In order to improve the security/stability of the MME, we want it to be robust against malformed user (UE) input. One method of doing this is fuzzing.

Here, we add a fuzzer that models corrupted S1AP packets and user controlled NAS packets.

NOTE: that fuzzing these right now essentially just fuzzes the S1AP decode logic and doesn't make it further as eNB has not been associated. This will be easier to fix once stateless MME has been finished as we can restore state from a snapshot with each fuzzing run instead of sending the right ITTI messages to "prime" the MME for fuzzing.

Further, the current test starts a "mock" SCTP task and the real S1AP task, which drops the packet. Once we have the correct S1AP state to not drop the packet, we also need to start MME and NAS task to receive the packet.

Reviewed By: ssanadhya

Differential Revision: D15695613

fbshipit-source-id: ad927a38f0a422bccd2ae24a02852b0c3a3582fe