Add pipelined and sessiond services to the CWAG
Summary:
This diff serves to add pipelined and sessiond services to
the CWAG. These services are necessary for the CWAG to function as
a PCEF. To comply with Magma's push to use docker, these services
are built with docker images, one for python (pipelined) and one
for C (sessiond). Since sessiond relies on a few custom packages,
the sessiond image is built using a base debian-stretch image.
Eventually, we should build any custom packages for xenial
to ensure that everything runs using the same base image.

The pipelined image also manually builds and installs OVS v2.8.1.
This is necessary for pipelined to use `ovs-vsctl` and `ovs-ofctl`.
To speed up installation in the future, we should host the `.deb.`
packages. Ths will speed up VM installation as well.

Lastly, this diff adds necessary service configs and OVS bridge
configuration to the CWAG ansible role. These may have to be
tweaked once further pipelined testing takes place.

Reviewed By: emakeev

Differential Revision: D15619491

fbshipit-source-id: e67536e72e99125e54e3322d27373cf558939845