Change feg docker services to run using bind networking (#345)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/345

Previously, the Federated Gateway's services could only run in host networking
mode. Due to Mac's lack of support for host networking, this caused some issues with
integration testing. This diff introduces a few changes to allow bind networking
while still remaining backwards compatible with Vagrant.

The main change this diff introduces is to make `service_registry.yml` a single source
of truth for the FeG's service registry. The main change needed for docker bind networking
is to set the ip_address used for each service as the service name (i.e. using magmad rather
than localhost). Rather than having two `service_registry` files (one for Vagrant and one
for Docker), the ansible role for the feg is updated to add localhost to /etc/hosts for each
service in that file.

Note: Since docker-compose doesn't yet support supplying sctp for a port, this will
have to be done manually using `docker-run -p ...`. If we want to use a script to do this,
we can add this in another diff.

Reviewed By: emakeev

Differential Revision: D15754801

fbshipit-source-id: 0225fa1410e0858b088846cba0c259222d1aacb1