Improve withAlert
Summary:
The generics on `withAlert` are wrong. We want it to take in a component that knows how to use the `WithAlert` type, pass those in internally in `withAlert`, and not require you to pass it externally.

Because of the incorrect types, many callsites used an empty object, resulting in no flow coverage on their components. This can be seen by the flow errors in those components (which I am not fixing because I am unsure on how to test).

I stumbled upon this when looking at D15528882 and thinking of a way to the `withAlert` use.

Reviewed By: murtadha

Differential Revision: D15762509

fbshipit-source-id: e538c27f8584560b6b51c306edc0657d58bde48e