Implementation of revalidation timer for RAR (#334)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/334

**Summary**
`session_proxy` in FeG might send `EventTriggers` and `RevalidationTime` to `sessiond` in the GW through RAR and CCU-A, but the `sessiond` does not have the implementation to trigger a CCU-R when the timer expires. This diff takes the information related to revalidation timer in RAR and implement the logic to trigger a CCU-R. There will be another diff doing the similar thing using CCU-A.

Some `using google::protobuf::RepeatedPtrField;` are removed as I noticed they are not needed.

**Implementation**
1. Get revalidation time from RAR message when `RevalidationTimeout` is one of the `EventTriggers`.
2. Schedule a rpc call to send CCU-R to `session_proxy` using EventBase.

**What is affected**
1. `sessiond` will send a CCU-R when revalidation timer it receives from RAR expires.

Reviewed By: emakeev

Differential Revision: D15707762

fbshipit-source-id: ff857eddae83eb05d8ec55175688c7b34b777d9a