Add REST api calls for test controller in orc8r (#328)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/328

Pull Request resolved: https://github.com/facebookexternal/fbc/pull/974

Create a test controller service that will allow for CI on the orc8r cloud. This commit creates a skeleton for the REST API calls for the controller and adds handlers for the api endpoints. It outlines functionality for automatic cloud upgrades based on a given package repo and target upgrade tier.

Reviewed By: xjtian

Differential Revision: D15719883

fbshipit-source-id: 10500c6e121882f24fcc3b94fdfff235e62764cf