Add ability to limit number of datapoints in prometheus cache
Summary: Request by Paul. You can specify the a limit when starting the process and if a metrics push comes in which would overfill the limit, it returns a 4xx status code and does not submit the metrics. Useful if you might push more often than prometheus pulls and don't want to use all of the system's memory

Reviewed By: xjtian

Differential Revision: D15870210

fbshipit-source-id: 827a092a2b6d4da63ac1f34f2bec113a281aeafe