Implement pipelined app to add subscriber flows using UE MAC addr (#351)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/351

This diff implements a pipelined app `ue_mac` to match packets based on
UE MAC address. This app controls table 0 and is necessary as the EPC app which
controls Table 0 for the LTE AGW is not present in the Carrier WiFi solution.

The app supports both adding and deleting flows based on MAC addr, but for now an
RPC only exists for the addition of a subscriber/mac addr. If necessary, an RPC can
easily be added to support deletion.

Differential Revision: D15873258

fbshipit-source-id: 988b0a9c382c22bba1cae68a2002ddd889a697c9