Attempt to make magma_get_config look similar to previous version
Summary:
The format of get_magma_config changed, which broke the seemingly only
usecase of it in the dns thing I wrote (which shouldn't break anything, just
not function).

It seems like SoMA is the only user of this script, so instead, let's make it simpler
to make it friendlier for bash parsing.

Namely:
1. Output one config per line (for piping)
2. Guarantee space separation (to allow a simple cut to get variables)

See test plan for details

Reviewed By: ilyacodesFB

Differential Revision: D15900550

fbshipit-source-id: 0981bfc22d24992830ae723254acd82a625ce466