Migrate cellular config REST handlers to use configurator (#367)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/367

The cellular configs API now supports the migration to configurator.  If "USE_NEW_HANDLERS" is set to 1, cellular configs will be backed by configurator, not configs.

I have duplicated the tests so that both versions are tested. The original tests are moved to cellular_handlers_legacy_test.go.

I've also modified the config type switch logic to look at the last index for the 'gateway' or 'network' keyword, not the second.

Reviewed By: xjtian

Differential Revision: D15907944

fbshipit-source-id: 9c8ae32acf6efd330d3dda92f02865783c2308ad