Switch config model validation to be done with swagger models not protobuf (#366)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/366

Since we are now switching over to storing swagger models instead of protobuf, I've moved the validation logic to be in terms of swagger models.
This will help when migrating the cellular handlers to use configurator, so that I can validate the inputs without converting into protobuf, which would require the now deprecated ToServiceModel function.

I also modified some of the error messages in the tests to match the new error messages.

Reviewed By: xjtian

Differential Revision: D15906242

fbshipit-source-id: 90eb8538cf57bf9ede3f495e34b93932b82ed4fc