Add string uppercase conversion to orc8r service registry map
Summary: When services are added to the registry in `service_registry.go` they are added with the name: `strings.ToUpper(name`). To ensure that services are accessed from the registry using the same name, this adds in a `strings.ToUpper()` call in the dictionary access. The original toUpper() appears to be due to legacy code from importing the service names as environment variables in `orc8r/cloud/docker/controller/Dockerfile`, so it may be worth looking into removing the need for upper case names or documenting that service names must be in all caps

Reviewed By: xjtian

Differential Revision: D15928240

fbshipit-source-id: 6b66e7728c74e96f7d8e62d891708d5349954645