Master org for system admin tools
Summary: We had an "admin" tab to host "system admin" tools, but that didn't really make sense since this shouldn't be configurable on the organization level. We want to be able to enable an admin tab for a "partner admin" tools, while the "system admin" is special cased as a "master org".

Reviewed By: noamsch

Differential Revision: D15963211

fbshipit-source-id: 2f0fb287aeaf7ace6a94152acbeeb7db62a433dc