Add radius, eap, aaa_server, redis to CWAG (#383)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/383

This diff serves to add RADIUS server, AAA server, EAP-AKA, and Redis services
to the CWAG. These are necessary for the completion of the CWAG and testing PCEF
functionality. Due to different deployment models, we use an env var to determine
where EAP-AKA should send an SWx request to.

Reviewed By: emakeev

Differential Revision: D15969191

fbshipit-source-id: 05f1dcba5f093890d1710aad88b72bef2afd17fb