Migrate Wifi Gateway Config to fit configurator association model (#1066)
Summary:
Pull Request resolved: https://github.com/facebookexternal/fbc/pull/1066

Pull Request resolved: https://github.com/facebookincubator/magma/pull/385

Since mesh and wifi gateway configs have a bidirectional dependency, this diff modifies the wifi gateway config logic to support this.
Each mesh has associations to gateways that it has and each gateway config has a meshID field that represents which mesh network the gateway is part of. So any time the gateway config is modified, the change should be reflected in the mesh entity associations field.

This feature can be turned on by the USE_NEW_HANDLERS env variable. Tests are duplicated to test both versions.

Reviewed By: xjtian

Differential Revision: D15972648

fbshipit-source-id: d536006e09f22d43af0b4a988dd1f237c2aaa2b3