Move docker magmad.yml to default feg configs
Summary:
Previously, we had a seperate directory for docker specific configs. This
is a bit risky given that they're not likely to be known and will go un-updated. This
diff moves these to the default configs directory at `magma/feg/gateway/configs`
and symlinks any legacy configs to ensure a Vagrant dev env still works.

Note: the NAT vagrantfile config is removed to ensure that FeG VM can still
connect to the orc8r. This was currently broken.

Reviewed By: emakeev

Differential Revision: D16043692

fbshipit-source-id: 1dbeab7cfa0acd48304a52dc1ed300c680ab1377