Add RADIUS layer on top of EAP-AKA support in UE SIM (#420)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/420

This adds UE-side Radius support on top of the EAP and EAP-AKA handlers.

Because there won't be a separate TWAN simulator, Radius handlers have been added to the UE Simulator.
Specifically, it wraps EAP packets -> Radius packets, and vice versa.
The verification step is missing right now because the simulator doesn't keep sessions or other state - that will have to come down the line.

Reviewed By: mpgermano

Differential Revision: D16094882

fbshipit-source-id: f9ba039756af1515506eb598b4fe46a278216ed6