Update RADIUS dockerfile to load configuration dynamically
Summary:
This diff corrects an issue that was caused by the radius dockerfile
not updating the radius config with the appropriate env-vars that are set.
This diff corrects this by using envsubst to do this before every run
inside the container.

This diff also updates the template config to reflect the fields of
`magma/feg/radius/src/radius.config.json` as this was outdated.

Reviewed By: themarwhal

Differential Revision: D16223474

fbshipit-source-id: df5f324e6020622334a143c179c70b5ed7ab365f