Correct exporter URLs on create
Summary: Graphite and Prometheus clients use different types of connections, so graphite needs to not have the protocol in the address, while prometheus does. Without correcting them it makes the configuration pretty fragile since the user would have to know about this when making the config file. Simply check for protocol and add/remove it from prometheus/graphite addresses respectively.

Reviewed By: tcirstea

Differential Revision: D16225640

fbshipit-source-id: ae08f35a7680ba07c23f8634d3ef647fbc734e02