Add rpc functionality to test controller service (#433)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/433

Pull Request resolved: https://github.com/facebookexternal/fbc/pull/1164

This provides a CLI script that can be run manually on a gateway with the command `./python/scripts/testcontroller_cli.py`

`testcontroller_cli.py` establishes two RPC connections. The first is with magmad on the gateway to query for the current magmad version. The second is with the testcontroller service in the cloud to report the magmad version and receive a script path in response. This path corresponds to a script on the gateway that will be executed, allowing for tests to be run automatically every time a gateway is upgraded to a new package version.

NOTE: Since the testcontroller service is in the fbinternal package, _pb2.py files are not generated automatically in the builds. Similar l to what is noted for vpnservice in D15200752, the _pb2.py files will need to be copied manually onto any gateway we wish to integrate with this setup. These proto files should be placed on the gateway in `/usr/local/lib/python3.5/dist-packages/` at the directory `fbinternal/protos`

NOTE: _pb2.py files will also have to be generated manually, which can be done using the command:
`python -m grpc_tools.protoc -I<src_path> --python_out=<dst_path> --grpc_python_out=<dst_path> <path_to_proto>`

Differential Revision: D16269581

fbshipit-source-id: aa2a4a6a385334ede71fa48cec69cebc30b212df