Add support for upgrading gateways through orc8r
Summary:
This diff enables proxying docker and github requests
through the orc8r. This is useful for production gateways that
cannot access any other parts of the internet. It works by
forwarding from nghttpx's 9444 to localhost 3128 where a
squid forward proxy is running. This is necessary as nghttpx
does not work as a forward proxy for anything other than
simple GET,POST forward proxying. Thus without squid,
HTTP CONNECT requests fail.

The squid configuration is the default for minimum access.

Reviewed By: amarpad

Differential Revision: D16403892

fbshipit-source-id: bf194611da4c2a2094cc6a9c85aeb032e9c883c4