Refactored generic health checking
Summary:
Refactored the generic health check functonality and splitted it up into several files. Added `magma/common/health` as a package.

Currently the Generic health checker is separate from the one specific for docker containers. Entities are in a separate file and the CLI is separate from the core logic.

Upgraded python-fire to use the latest fancy features.

Reviewed By: ssanadhya

Differential Revision: D16292388

fbshipit-source-id: 42e4690c67201cae98a36e64eaf95ef1f5646d77