Add configurable option to exclude Destination-Host AVP in diameter requests
Summary:
Recently, we've encountered the need for Destination-Host to be configurably
included or excluded in Diameter requests sent from the FeG. This option is in line with the
specs of SWx, S6a, Gx, and Gy. This diff implements this functionality by adding an mconfig
parameter to the DiameterServerConfig. During the function where the dest-host and dest-realm
AVPs are added to outgoing diameter requests, we first check this configuration before adding
dest-host.

Reviewed By: andreilee

Differential Revision: D16504258

fbshipit-source-id: f9acc7949c348ad68eafaee7cd6cb83f0b733142