Add a type field to network for classification
Summary:
Add a type field to network to enable typed networks for API v1. Since this depends on the cfg_networks table to have a type column, the migration004 script must be run in /var/opt/magma/bin for configurator to work properly.

Also extended the new type field to the configurator client API.

Reviewed By: xjtian

Differential Revision: D16349740

fbshipit-source-id: 38013dba3396e4642d3c7e21c5a1a469cc214dc2