Add a type field to network for classification (#484)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/484

Add a type field to network to enable typed networks for API v1. Since this depends on the cfg_networks table to have a type column, the migration004 script must be run in /var/opt/magma/bin for configurator to work properly.

Also extended the new type field to the configurator client API.

Reviewed By: xjtian

Differential Revision: D16349740

fbshipit-source-id: 86ae48e293bec0121b16614433c3e8a8a6797a16