Migrate dnsd network configs model and network features model usages to point to new models defined under pluginimpl
Summary:
Changing the network config serdes and other references of NetworkDNSConfig and NetworkFeatures to point to the models defined in pluginimpl/models.
The corresponding models defined in services/dns and services/magmad should be deprecated when magma-v0 is deprecated.

I'm only touching the two models in this diff but others will have to be migrated similarly as well. Will do that in following diffs.

Reviewed By: xjtian

Differential Revision: D16552186

fbshipit-source-id: cd56cf2cae793c02e2cd13fcaa156fc9a869fcec