Add testcontroller docker-compose override file
Summary:
For the test controller service to execute scripts on the testcontroller gateway, the ssh keys and the scripts need to be mounted into the docker container. This mounts those as volumes.
magmad docker container should be brought up with `docker-compose -f docker-compose.yml -f docker-compose.override.yml -f docker-compose.testcontroller.yml up -d`

ssh and fabric are needed to run the scripts on the testcontroller gateway, so this modifies the Dockerfile to include those dependencies

Reviewed By: mpgermano

Differential Revision: D16592401

fbshipit-source-id: 1711dcb5d0c7112c47848934bb1bbc22a8b4536b