Build cache dir before building cache, and check return code
Summary:
I had an inconsistent state due to the cache failing to build, but the error
was masked and the *next* build was failed.  We should bail if the cache fails
to build.

Also, I was trying to fix the broken state, but the tmp directory wasn't being updated,
so I would make a change but not see it reflected in the build

Reviewed By: xjtian

Differential Revision: D16631263

fbshipit-source-id: d3ce30a9593e2db2f7a2eeba2f8c3604b31f24b5