Implement the rest of the radiusd service (#525)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/525

This revision implements the actual functionality of the radiusd service.

- Fetches the prometheus metrics from the radius server
- Processes the metrics and re-exports them to orc8r

To read in metrics exported from the Radius server, and re-export them to orc8r, we've created the MetricAggregate interface. This handles reading in the MetricFamily exported from Radius, and then updating the corresponding collector in radiusd's prometheus registry.

A wrapper is created for the prometheus default registry, in registry_wrapper.go. The single Update function takes in the exported metrics from Radius, and will create and register new metrics, and prevent repeat registrations.

Reviewed By: xjtian

Differential Revision: D16286847

fbshipit-source-id: 4743452af678a204b00a31d720757e75a1a4c466