[ltel Add generic command for rebooting eNB from NMS (#519)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/519

This revision adds the ability for remote commands to be run to have the Access Gateway reboot any and all connected eNodeB devices.

This is intended to help with ease of use, to allow users to execute some basic troubleshooting functionality through a UI rather than through the command line.

Reviewed By: xjtian

Differential Revision: D16473195

fbshipit-source-id: 54153d7359bc7387be66d6d3a70707c5d720a86d