Fix sctp_bindx error when Sctpd restarts
Summary:
The sctp socket does not get released immediately after the sctpd process dies.
Thus, when sctpd restarts, the new process is unable to create a new sctp socket
on the same port. This change fixes the socket options, so that SCTP ABORT is
sent whenever the process dies, and the socket is released immediately.

Reviewed By: shahrahman-fb

Differential Revision: D16648938

fbshipit-source-id: 92a9dc10ae9c5021d014dd48d29b2d3b40c4db4a