Add support for HSS initiated de-registration for SWx (#532)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/532

This diff adds support for Registration-Termination to both the HSS
and SWx Proxy services. This diameter command requests to the AAA server to kick
the provided subscriber off of the network. In order to support this functinality,
the HSS must keep a mapping of clients and their IP addresses. It should re-use
an existing connection if the diameter connection already exists, otherwise
the request will not succeed.

In order to more easily test this functionality, an interface for the SWx proxy to
relay is extracted to allow it to be mocked out. Support for the call from the
SWx proxy service to FeG-Relay will be added in a later diff.

Reviewed By: xjtian

Differential Revision: D16662589

fbshipit-source-id: 0db6e2adbd9483268dcba265f9dce309551eae91