withStyles fixes
Summary:
It seems like the style of `theme.spacing.unit * 3` is still occasionally being added - this is deprecated.
This diff fixes it for the use of `withStyles`. It doesn't fix it for `makeStyles` since that also accepts using props in the css which I can't figure out how to type properly.

The reason this diff works is because "StyleRules is actually a general enough type that it is a supertype of functions. Remember, functions are objects too. So, you want Flow to notice that styles is a function type, and typecheck it against StyleRulesCallback"
More can be read here: https://fb.workplace.com/groups/flow/permalink/2680931811955417/

I also flipped the order of StyleRules and StyleRulesCallback because this will cause flow errors if people change it incorrectly.

Reviewed By: rckclmbr

Differential Revision: D16707031

fbshipit-source-id: ad21bd8de198b2c64b66afbbbb927171383756a4