TR-069 sessions with unregistered devices will be rejected
Summary: This revision changes the behavior of enodebd service to reject TR-069 sessions initiated by eNB devices that have not been registered to the Access Gateway, if at least one is registered. This is a step in fully removing support for the old method of configuring eNB. Unregistered eNB devices should not be configured by the Access Gateway and should not work.

Reviewed By: fishlinghu

Differential Revision: D16744857

fbshipit-source-id: 68ee3b859ec015368f3c40541ebc92c612ca369d