TR-069 sessions with unregistered devices will be rejected (#540)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/540

This revision changes the behavior of enodebd service to reject TR-069 sessions initiated by eNB devices that have not been registered to the Access Gateway, if at least one is registered. This is a step in fully removing support for the old method of configuring eNB. Unregistered eNB devices should not be configured by the Access Gateway and should not work.

Reviewed By: fishlinghu

Differential Revision: D16744857

fbshipit-source-id: 96db1b3d6a63472df06af1103c811570d63f2a7e