Fix fab register_vm command to use the new gateway config
Summary: The lte gateway fab command for registering a gateway was still using the old gateway config, resulting in an error `Exception: Received a 400 response: {"message":"Invalid Gateway Record, Error: validation failure list:\nhardware_id in body is required"}`. This changes it so that it's able to properly register a gw.

Differential Revision: D16779947

fbshipit-source-id: b5e2823b3198ad62ef3c83f3b1271091c1bd744e