Fix fab register_vm command to use the new gateway config (#545)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/545

The lte gateway fab command for registering a gateway was still using the old gateway config, resulting in an error `Exception: Received a 400 response: {"message":"Invalid Gateway Record, Error: validation failure list:\nhardware_id in body is required"}`. This changes it so that it's able to properly register a gw.

Reviewed By: xjtian

Differential Revision: D16779947

fbshipit-source-id: 69d87aa961b06a1a74e4a2df3bce688cc491c624