Add a configurable tolerance margin for extra quota
Summary:
**Summary**
This diff changes the logic for deciding when to deactivate the service (stop the data).

__Old logic:__
There were two situations when the service would be stopped:

-
```
Pipelined reports data usage
  --> quota is exhausted
  --> sessiond sends update to FeG, hoping to get successful response from core and have more quota granted
  --> get unsuccessful response (still a successful GRPC to FeG)
```
-
```
Pipelined reports data usage
  --> quota is exhausted, and it is the final grant
```
Here was one strange case. sessiond would continue the service if the GRPC to FeG for updating credit usage failed, no matter how much data a session already used.

__New logic:__
First, two new configurable parameters are introduced:
1. `terminate_service_when_quota_exhausted`: Set to `false` if we don't want to terminate a session no matter what. Otherwise, set to `true`.
2. `extra_quota_margin`: Extra number of bytes a session could use after the quota is exhausted. This is like a tolerance margin. After a session used up `(granted quota + extra tolerance quota)`, the service would be deactivated.

Now, there is a new case when the service would be stopped.

-
```
Pipelined reports data usage
  --> quota is exhausted, and the extra quota is exhausted as well
```
Therefore, a session will no longer be able to use data forever when the GRPC for updating credit usage continues to fail. The service will be deactivated when the quota is exhausted no matter what, but we still allow a configurable tolerance margin to make sure the service could continue while sessiond is trying to have a successful GRPC.

**Implementation**
1. Add configurable parameters in `sessiond.yml`
2. Change the logic for deciding when to deactivate the service
3. Update unit test

**What is affected**
Service would be deactivated when the quota is exhausted.

Reviewed By: emakeev

Differential Revision: D16699719

fbshipit-source-id: 149f3225b4780546bdf0d3b96389728b4d55c9b5