Get metricsd running on cloud proxy tier (#555)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/555

This revision only makes sure that metricsd can run on the proxy tier, but does not ensure that the correct metrics are being collected or exported with the correct labels.

The `Dockerfile` for cloud proxy tier has been updated to include a lightweight build of orc8r with no other modules/plugins.

`supervisord.conf` has been modified so that supervisord will now run metricsd on the proxy tier, as well as the supervisor_logger.

Reviewed By: xjtian

Differential Revision: D16372814

fbshipit-source-id: 4ddee511bd199fbc17218af1b4efd4fc76fd27ed