Add docker upgrader to magmad (#567)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/567

This diff serves to add functionality to magmad to allow for
orc8r initiated upgrades of docker-based containers. The functionality
works by parsing a version "0.3.77-1565647911-e53c1fe2" => e53c1fe2
and then using that tag as the IMAGE_VERSION. The upgrader then clones
the github repo, checks out that build tag and updates the docker-compose
file.

Since a docker service cannot recreate itself with a new image, an external
process must do this. In this case, we just use a cron job checking every 15
minutes to see if the currently running tag and the updated tag differ.
Though this leaves a small time period with an unfinished update, if a service
were to restart for any reason, the docker-compose changes won't be included,
as docker doesn't update config settings on restart. This ensuresthat we cannot
get into a bad state during an upgrade.

Reviewed By: andreilee

Differential Revision: D16811961

fbshipit-source-id: d867f9c247518016d78be5149780bb6654f24312