Fix staging configurator crash due to imcompatible types
Summary:
The serde was migrated to use the new GatewayCellularConfigs defined in plugin/models while the mconfig was still trying to interpret it as the old model defined in cellular/obsidian/models. The unit test did not catch this because it was not testing the de/serialization and was using the old models.

Updated the test to use the new models.

Differential Revision: D16918313

fbshipit-source-id: 571f867d2c2e94d43b389238781b341bd49f72b9