Fix staging configurator crash due to imcompatible types (#573)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/573

The serde was migrated to deserialize into the new GatewayCellularConfigs defined in plugin/models while the mconfig was still trying to interpret it as the old model defined in cellular/obsidian/models.
```panic: interface conversion: interface {} is *models.GatewayCellularConfigs, not *models.GatewayCellularConfigs (types from different packages)```

The unit test did not catch this because it was not testing the de/serialization and was using the old models.

Updated the test to use the new models.

Reviewed By: mpgermano

Differential Revision: D16918313

fbshipit-source-id: 6aa366686ce296888c6b0d6c9562d1e3e4ff96b4