Finish migrating from magmad to configurator for nw/gw lookup and reduce error logs
Summary:
Remove the conditional on env vars because it should be always reading from configurator now.
When I was putting the migrating logic in, I introduced more error logs than necessary but now checking for empty hwid before looking up networkID and gatewayID, leading to a bunch of Empty HWID lookup error printed out in the cloud. So fixing that here.

Differential Revision: D16939909

fbshipit-source-id: 4402b239af4af82591e556a8d61e763270adc761