Finish migrating from magmad to configurator for nw/gw lookup and reduce error logs (#579)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/579

Remove the conditional on env vars because it should be always reading from configurator now.
When I was putting the migrating logic in, I introduced more error logs than necessary but now checking for empty hwid before looking up networkID and gatewayID, leading to a bunch of Empty HWID lookup error printed out in the cloud. So fixing that here.

Reviewed By: xjtian

Differential Revision: D16939909

fbshipit-source-id: b2e651b224e3078b4fc92c64173628e674c8c274