Add a PartialGatewayRead and PartialGatewayUpdate framework to have a generic way to modify gateways. (#580)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/580

Added the following to have a generic way to modify all entity components that implement PartialEntityModels as well as reading the physical device/state attached to the entity.
The setup is relatively similar to the PartialNetwork Update/Read framework. The main difference is that we pass in paramName to grab the appropriate entityKey, and the entityType.

- `GetPartialReadGatewayHandler(path string, model PartialGatewayModel) obsidian.Handler`
- `GetPartialUpdateGatewayHandler(path string, model PartialGatewayModel) obsidian.Handler`
- `GetReadGatewayStateHandler(path string) obsidian.Handler`
- `GetReadGatewayDeviceHandler(path string) obsidian.Handler`

Reviewed By: xjtian

Differential Revision: D16912731

fbshipit-source-id: 9bc43349850a3a698a82ae72140f0c26bc3b2114